gitrepo/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
importlib-metadata==4.12.0
lxml==4.9.1
Markdown==3.4.1
oauthlib==3.2.0
postgres==4.0
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
pytz==2022.2
requests==2.28.1
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.11
zipp==3.8.1
